Se crea el método de menú
